+++ src/components/atoms/Button/index.stories.tsx
@@ -1,0 +1,16 @@
+import { ComponentStory, ComponentMeta } from '@storybook/react';
+
+import { Button } from './index';
+
+export default {
+  title: 'Example/atoms/Button',
+  component: Button,
+  argTypes: { onClick: { action: 'clicked' } },
+} as ComponentMeta<typeof Button>;
+
+const Template: ComponentStory<typeof Button> = (args) => <Button {...args} />;
+
+export const Default = Template.bind({});
+Default.args = {
+  value: 'button',
+};
